# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  main.JSONResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
  main.TokenPairs:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample contacts server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OWT Swagger API
  version: "1.0"
paths:
  /authenticate:
    post:
      consumes:
      - application/json
      description: authenticate a user with email & password
      parameters:
      - description: email
        in: query
        maxLength: 255
        minLength: 5
        name: string
        type: string
      - description: password
        in: query
        maxLength: 255
        minLength: 1
        name: string
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/main.TokenPairs'
              type: object
      summary: Authentication
      tags:
      - Tools
  /contact/{id}:
    get:
      consumes:
      - application/json
      description: get a contact with id, token
      parameters:
      - description: contact id
        in: query
        maximum: 65535
        minimum: 0
        name: int
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.JSONResponse'
      summary: GetContact
      tags:
      - Tools
  /contact/all:
    get:
      consumes:
      - application/json
      description: list all contacts with token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.JSONResponse'
      summary: AllContacts
      tags:
      - Tools
  /contact/new:
    put:
      consumes:
      - application/json
      description: create a new contact with first_name, last_name, full_name, email,
        password, address, mobile, token
      parameters:
      - description: first_name
        in: query
        maxLength: 20
        minLength: 1
        name: string
        type: string
      - description: last_name
        in: query
        maxLength: 20
        minLength: 1
        name: string
        type: string
      - description: full_name
        in: query
        maxLength: 40
        minLength: 3
        name: string
        type: string
      - description: email
        in: query
        maxLength: 255
        minLength: 5
        name: string
        type: string
      - description: password
        in: query
        maxLength: 255
        minLength: 1
        name: string
        type: string
      - description: address
        in: query
        maxLength: 255
        minLength: 1
        name: string
        type: string
      - description: phone number
        in: query
        maxLength: 20
        minLength: 1
        name: string
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.JSONResponse'
            - properties:
                data:
                  type: integer
              type: object
      summary: InsertContact
      tags:
      - Tools
  /contact/update/{id}:
    patch:
      consumes:
      - application/json
      description: update a contact by id with content and skills, token
      parameters:
      - description: contact id
        in: query
        maximum: 65535
        minimum: 0
        name: int
        type: integer
      - description: first_name
        in: query
        maxLength: 20
        minLength: 1
        name: string
        type: string
      - description: last_name
        in: query
        maxLength: 20
        minLength: 1
        name: string
        type: string
      - description: full_name
        in: query
        maxLength: 40
        minLength: 3
        name: string
        type: string
      - description: email
        in: query
        maxLength: 255
        minLength: 5
        name: string
        type: string
      - description: address
        in: query
        maxLength: 255
        minLength: 1
        name: string
        type: string
      - description: phone number
        in: query
        maxLength: 20
        minLength: 1
        name: string
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.JSONResponse'
            - properties:
                message:
                  type: string
              type: object
      summary: UpdateContact
      tags:
      - Tools
  /delete/{id}:
    get:
      consumes:
      - application/json
      description: delete a contact with id, token
      parameters:
      - description: contact id
        in: query
        maximum: 65535
        minimum: 0
        name: int
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.JSONResponse'
            - properties:
                message:
                  type: string
              type: object
      summary: DeleteContact
      tags:
      - Tools
  /skill/{id}:
    delete:
      consumes:
      - application/json
      description: delete a skill with id, token
      parameters:
      - description: skill id
        in: query
        maximum: 65535
        minimum: 0
        name: int
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.JSONResponse'
            - properties:
                message:
                  type: string
              type: object
      summary: DeleteSkill
      tags:
      - Tools
    get:
      consumes:
      - application/json
      description: get a skill with id, token
      parameters:
      - description: skill id
        in: query
        maximum: 65535
        minimum: 0
        name: int
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.JSONResponse'
      summary: GetSkill
      tags:
      - Tools
  /skill/all:
    get:
      consumes:
      - application/json
      description: list all skills with token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.JSONResponse'
      summary: AllSkills
      tags:
      - Tools
  /skill/new:
    put:
      consumes:
      - application/json
      description: create a new skill with name, level, token
      parameters:
      - description: name
        in: query
        maxLength: 20
        minLength: 1
        name: string
        type: string
      - description: level
        in: query
        maximum: 4
        minimum: 0
        name: int
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.JSONResponse'
            - properties:
                data:
                  type: integer
              type: object
      summary: InsertSkill
      tags:
      - Tools
  /skill/update/{id}:
    patch:
      consumes:
      - application/json
      description: update a skill by id with with name, level, token
      parameters:
      - description: skill id
        in: query
        maximum: 65535
        minimum: 0
        name: int
        type: integer
      - description: level
        in: query
        maximum: 4
        minimum: 0
        name: int
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.JSONResponse'
            - properties:
                message:
                  type: string
              type: object
      summary: UpdateSkill
      tags:
      - Tools
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"